server/requirements.txt
alembic==1.4.2
appnope==0.1.0
backcall==0.2.0
bleach==1.5.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
decorator==4.4.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gevent==20.6.2
geventhttpclient==1.4.4
greenlet==0.4.16
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==1.7.0
ipython
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi
Jinja2==2.11.2
Keras==2.1.2
locust==1.1.1
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.19.0
opencv-python==4.3.0.36
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.2
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
scipy==1.5.1
six==1.15.0
SQLAlchemy==1.3.18
tensorflow-tensorboard==0.1.8
traitlets==4.3.3
urllib3==1.25.9
wcwidth
Werkzeug==1.0.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
